+++ app/controllers/services_controller.rb
@@ -11,10 +11,15 @@ class ServicesController < ApplicationController
   end
 
   def show
-    @service = Service.find(params[:id])
+    @service = Service.
+               includes(:offers, related_services: :provider).
+               find(params[:id])
+
+    @offers = @service.offers
+    @related_services = @service.related_services
+
     @service_opinions = ServiceOpinion.joins(project_item: :offer).
                         where(offers: { service_id: @service })
     @question = Service::Question.new(service: @service)
-    @related_services = @service.related_services.includes(:provider)
   end
 end