+++ CHANGELOG.md
@@ -78,6 +78,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 - Sidekiq for delayed jobs (@mkasztelnik)
 - Admin administration panel stub with sidekiq monitoring and MP version (@mkasztelnik)
 - Render root categories in page footer (@mkasztelnik)
+- Use bootstrap cards to show service offers (@mkasztelnik)
 
 ### Changed
 - Upgrade Sprockets gem to avoid CVE-2018-3760 vulnerability (@mkasztelnik)